Phase 3.3: unified ProductHeader + sticky compact mode
Replaces the per-page ProductHeaderBar across /research/[id],
/vetting/[asin], /offer/[id], /sourcing/[asin] with a single
component that adds funnel-stage progress + a sticky compact
variant on scroll.

- New ProductHeader (src/components/Product/ProductHeader.tsx)
  carries forward everything ProductHeaderBar did (image, alias,
  pencil rename, ASIN link, left/right action buttons, optional
  badge, optional inline action) and adds two new layers:
    1. Stage strip — Research → Vetting → Offering → Sourcing
       chips below the title, lit per data presence, current
       phase emphasized, lit non-current chips link to the
       matching detail page.
    2. Sticky compact mode — IntersectionObserver sentinel
       toggles a data-sticky attribute when the header pins
       below the global AppHeader (top: 64px). Compact layout
       collapses to thumb + name + current-stage pill + the
       primary right action only. Instant render swap, no
       transition.
- Stage data is computed at each call site from data already in
  scope (no extra fetches): submission.score for vetted,
  research_products.is_offered / is_sourced for offered/sourced.
- ProductHeaderBar.tsx deleted in this commit — only the four
  detail pages imported it and all four were migrated.
